+++ examples/create_customer_with_card.rs
@@ -1,10 +1,6 @@
-extern crate stripe;
-
-use std::env;
-
 fn main() {
     // Create a new client
-    let secret_key = env::var("STRIPE_SECRET_KEY").expect("Missing STRIPE_SECRET_KEY in env");
+    let secret_key = std::env::var("STRIPE_SECRET_KEY").expect("Missing STRIPE_SECRET_KEY in env");
     let client = stripe::Client::new(secret_key);
 
     // Define a card for the customer
@@ -30,9 +26,6 @@ fn main() {
     email: {:?},
     ..
 }}",
-        customer.id,
-        customer.created,
-        customer.default_source,
-        customer.email
+        customer.id, customer.created, customer.default_source, customer.email
     );
 }