+++ src/async.rs
@@ -1,21 +1,22 @@
 use crate::config::Response;
 use crate::error::{Error, ErrorResponse, RequestError};
 use crate::params::Headers;
+use futures::future::{self, Future};
+use futures::stream::Stream;
 use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
-use reqwest::RequestBuilder;
+use reqwest::r#async::RequestBuilder;
 use serde::de::DeserializeOwned;
-use std::io::Read;
 
 #[derive(Clone)]
 pub struct Client {
-    client: reqwest::Client,
+    client: reqwest::r#async::Client,
     secret_key: String,
     headers: Headers,
 }
 
 impl Client {
     pub fn new<S: Into<String>>(secret_key: S) -> Client {
-        let client = reqwest::Client::new();
+        let client = reqwest::r#async::Client::new();
         Client { client, secret_key: secret_key.into(), headers: Headers::default() }
     }
 
@@ -38,57 +39,66 @@ impl Client {
     }
 
     /// Make a `GET` http request with just a path
-    pub fn get<T: DeserializeOwned>(&self, path: &str) -> Response<T> {
+    pub fn get<T: DeserializeOwned + Send + 'static>(&self, path: &str) -> Response<T> {
         let url = Client::url(path);
         let request = self.client.get(&url).headers(self.headers());
         send(request)
     }
 
     /// Make a `GET` http request with url query parameters
-    pub fn get_query<T: DeserializeOwned, P: serde::Serialize>(
+    pub fn get_query<T: DeserializeOwned + Send + 'static, P: serde::Serialize>(
         &self,
         path: &str,
         params: P,
     ) -> Response<T> {
-        let url = Client::url_with_params(path, params)?;
+        let url = match Client::url_with_params(path, params) {
+            Err(err) => return Box::new(future::err(err)),
+            Ok(ok) => ok,
+        };
         let request = self.client.get(&url).headers(self.headers());
         send(request)
     }
 
     /// Make a `DELETE` http request with just a path
-    pub fn delete<T: DeserializeOwned>(&self, path: &str) -> Response<T> {
+    pub fn delete<T: DeserializeOwned + Send + 'static>(&self, path: &str) -> Response<T> {
         let url = Client::url(path);
         let request = self.client.delete(&url).headers(self.headers());
         send(request)
     }
 
     /// Make a `DELETE` http request with url query parameters
-    pub fn delete_query<T: DeserializeOwned, P: serde::Serialize>(
+    pub fn delete_query<T: DeserializeOwned + Send + 'static, P: serde::Serialize>(
         &self,
         path: &str,
         params: P,
     ) -> Response<T> {
-        let url = Client::url_with_params(path, params)?;
+        let url = match Client::url_with_params(path, params) {
+            Err(err) => return Box::new(future::err(err)),
+            Ok(ok) => ok,
+        };
         let request = self.client.delete(&url).headers(self.headers());
         send(request)
     }
 
     /// Make a `POST` http request with just a path
-    pub fn post<T: DeserializeOwned>(&self, path: &str) -> Response<T> {
+    pub fn post<T: DeserializeOwned + Send + 'static>(&self, path: &str) -> Response<T> {
         let url = Client::url(path);
         let request = self.client.post(&url).headers(self.headers());
         send(request)
     }
 
     /// Make a `POST` http request with urlencoded body
-    pub fn post_form<T: DeserializeOwned, F: serde::Serialize>(
+    pub fn post_form<T: DeserializeOwned + Send + 'static, F: serde::Serialize>(
         &self,
         path: &str,
         form: F,
     ) -> Response<T> {
         let url = Client::url(path);
         let request = self.client.post(&url).headers(self.headers());
-        let request = with_form_urlencoded(request, &form)?;
+        let request = match with_form_urlencoded(request, &form) {
+            Err(err) => return Box::new(future::err(err)),
+            Ok(ok) => ok,
+        };
         send(request)
     }
 
@@ -136,57 +146,25 @@ fn with_form_urlencoded<T: serde::Serialize>(
         .body(body))
 }
 
-fn send<T: DeserializeOwned>(request: RequestBuilder) -> Response<T> {
-    let mut response = request.send()?;
-    let mut body = String::with_capacity(4096);
-    response.read_to_string(&mut body)?;
+fn send<T: DeserializeOwned + Send + 'static>(request: RequestBuilder) -> Response<T> {
+    Box::new(request.send().map_err(Error::Http).and_then(|response| {
+        let status = response.status();
+        response.into_body().concat2().map_err(Error::Http).and_then(move |body| {
+            let mut body = std::io::Cursor::new(body);
+            let mut bytes = Vec::with_capacity(4096);
+            std::io::copy(&mut body, &mut bytes)?;
 
-    let status = response.status();
-    if !status.is_success() {
-        let mut err = serde_json::from_str(&body).unwrap_or_else(|err| {
-            let mut req = ErrorResponse { error: RequestError::default() };
-            req.error.message = Some(format!("failed to deserialize error: {}", err));
-            req
-        });
-        err.error.http_status = status.as_u16();
-        return Err(Error::from(err.error));
-    }
+            if !status.is_success() {
+                let mut err = serde_json::from_slice(&bytes).unwrap_or_else(|err| {
+                    let mut req = ErrorResponse { error: RequestError::default() };
+                    req.error.message = Some(format!("failed to deserialize error: {}", err));
+                    req
+                });
+                err.error.http_status = status.as_u16();
+                return Err(Error::from(err.error));
+            }
 
-    serde_json::from_str(&body).map_err(Error::deserialize)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{with_form_urlencoded, Client};
-    use crate::CustomerParams;
-    use std::collections::HashMap;
-
-    #[test]
-    fn serialize_metadata() {
-        let mut metadata = HashMap::new();
-        metadata.insert("any".to_string(), "thing".to_string());
-        let form = CustomerParams {
-            email: Some("jdoe@example.org"),
-            metadata: Some(metadata),
-            // ...
-            source: None,
-            default_source: None,
-            account_balance: None,
-            business_vat_id: None,
-            coupon: None,
-            description: None,
-            shipping: None,
-        };
-        let url = Client::url("/");
-        let http = reqwest::Client::new();
-        let result = with_form_urlencoded(http.post(&url), &form).and_then(|x| Ok(x.build()?));
-        assert!(result.is_ok(), "Failed to build request: {:?}", result);
-        if let Ok(request) = result {
-            let body = format!("{:?}", request.body());
-            assert_eq!(
-                body,
-                "Some(Body { kind: b\"email=jdoe%40example.org&metadata[any]=thing\" })"
-            );
-        }
-    }
+            serde_json::from_slice(&bytes).map_err(Error::deserialize)
+        })
+    }))
 }